chore(tests): update test environment
- bump vite from 4.3.9 to 4.4.4
- bump vitest from 0.32.2 to 0.33.0
- drop @vitest/ui and serve
- reorganize fixtures
- replace @vitest/coverage-c8 with @vitest/coverage-v8
- https://github.com/vitest-dev/vitest/pull/3339

+++ __fixtures__/float.ts
@@ -3,6 +3,7 @@
  * @module fixtures/FLOAT
  */
 
+import type { Float } from '#src/types'
 import INT from './integer'
 
-export default INT / 100
+export default (INT / 100) as Float
